Update how listings are got and deleted
* Use specific ListingService function to get the user's listings, instead of iterate through all listings from all users and checking which ones has the logged in user id;
* Before deleting a listing, check if  the user deleting it is the user which created it;
Also:
* Update the form that shows a listing to make the "Back" button go back to the user's account page and not to the edition or creation page;

+++ src/main/java/com/neon/controller/ListingController.java
@@ -56,16 +56,11 @@ public class ListingController {
 
     @GetMapping("/{id}")
     public String showListing(Model model, @PathVariable("id") Integer id){
-        if (id != null) {
-            Listing listing = listingService.findOne(id).get();
-            if (listing != null){
-                User seller = userService.findOne(listing.getSellerId()).get();
-                model.addAttribute("user", seller);
-                model.addAttribute("listing", listing);
-                return "/listing/show";
-            }
-        }
-        return "/";
+        Listing listing = listingService.findOne(id).get();
+        User seller = userService.findOne(listing.getSellerId()).get();
+        model.addAttribute("user", seller);
+        model.addAttribute("listing", listing);
+        return "/listing/show";
     }
 
     @GetMapping("/search")
@@ -86,8 +81,12 @@ public class ListingController {
     @GetMapping("/{id}/delete")
     public String deleteListing(Model model, @PathVariable("id") Integer listingId) {
         Listing listing = listingService.findOne(listingId).get();
-        Integer userID = listing.getSellerId();
-        listingService.delete(listing);
-        return "redirect:/users/" + userID;
+        Integer sellerId = listing.getSellerId();
+        String loggedInUsername = SecurityUtils.getLoggedInUsername();
+        User user = userService.findOneByUsername(loggedInUsername);
+        if (sellerId == user.getId()) {
+            listingService.delete(listing);
+        }
+        return "redirect:/users/my-account";
     }
 }
